Rename dual_n_channel_mosfet_transistor to ground_bridge
