+++ symbols/ground_bridge_left.ts
@@ -1,0 +1,4 @@
+import { flipSymbolOverYAxis } from "drawing/rotateSymbol"
+import ground_bridge_right from "./ground_bridge_right"
+
+export default flipSymbolOverYAxis(ground_bridge_right)
